week_06/day_1/types_start_point/src/main/java/Bear.java
public class Bear {
    private String name;
    private int age;
    private double weight;
    private String address;

    public Bear(String name, int age, double weight){
        this.name = name;
        this.age = age;
        this.weight = weight;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }

    public double getWeight() {
        return weight;
    }

    public void setWeight(double weight) {
        this.weight = weight;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public boolean readyToHibernate() {
//        if (this.weight > 80.0) {
//            return true;
//        }
        return this.weight > 80.0;
    }
}
